moved mapping functions to common file, fixed major bug with data sources but still WIP
