+++ examples/aks_cluster/versions.tf
@@ -2,11 +2,11 @@ terraform {
   required_providers {
     azurerm = {
       source  = "hashicorp/azurerm"
-      version = "2.66.0"
+      version = ">= 3.27.0"
     }
     tsm = {
       source  = "robwafle/tsm"
-      version = "~> 0.0.1"
+      version = ">= 0.0.81"
     }
     # cloudflare = {
     #   source  = "cloudflare/cloudflare"
@@ -21,6 +21,5 @@ terraform {
     #   version = "=2.8.0"
     # }
   }
-  required_version = ">= 0.14"
 }
