+++ examples/global_namespace/providers.tf
@@ -1,5 +1,3 @@
 provider "tsm" {
-  //host = "https://prod-4.nsxservicemesh.vmware.com"
-  //apikey = ""
 }
